Separate CustomSystem* from its database
